# Base64 encode SVG