--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "guitar-tuner",
-  "version": "0.0.66",
+  "version": "0.0.68",
   "private": true,
   "engines": {
     "node": ">=0.10.0"
@@ -28,6 +28,7 @@
     "gulp-bump": "^0.3.0",
     "gulp-replace": "^0.5.3",
     "gulp-minify-inline": "^0.1.0",
+    "gulp-base64": "^0.1.3",
     "gulp-vulcanize": "^6.0.0"
   }
 }