Web Page Performance Test for https://guitar-tuner.appspot.com
From: Dulles, VA - Motorola G - Chrome - 3G

Before: First view Visually Complete: 9.913s
After: First view Visually Complete: 7.893s

Before: Repeat view Visually Complete: 4.087s
After: Repeat view Visually Complete: 4.142s

http://www.webpagetest.org/result/150703_AX_KJ8/1/details/cached/
http://www.webpagetest.org/result/150703_J6_M0S/1/details/cached/