@beaufortfrancois 2015-07-24
Conflicts have been fixed @paullewis 
Sorry for that!