@addyosmani 2016-03-11
@paullewis This should be good to go afaik. Might want to ask for commits to be squashed for a cleaner merge history, but otherwise should be fine.